P2a: backend runtime compat for torch 2.13 / diffusers 0.39 / transformers 5 / hf_hub 1.x / Pillow 12 / numpy 2
- torch.load weights_only=True at all 24 call sites
- from_single_file: original_config rename, explicit safety-checker components preserving both NSFW branches (+ regression test)
- typed huggingface_hub exceptions replacing string matching
- transformers 5: explicit dtype on all loads; CPUTextEncoderWrapper eager attn config; briarmbg2 model_type/tied-weights compat
- vendored SAM2: drop deprecated restrictive sdp_kernel contexts (fixes Blackwell)
- rembg cloth_seg stacked-mask normalization in RemoveBG.gen_mask
- AnyText mechanical fixes: getsize/getoffset -> getbbox, np.int0 -> np.intp
- imghdr (removed in py3.13) -> Pillow-based sniffing
- gradio lazy import for web-config extra; ldm autocast deprecation

+++ iopaint/model/anytext/ldm/modules/encoders/modules.py
@@ -96,7 +96,9 @@ class FrozenT5Embedder(AbstractEncoder):
     ):  # others are google/t5-v1_1-xl and google/t5-v1_1-xxl
         super().__init__()
         self.tokenizer = T5Tokenizer.from_pretrained(version)
-        self.transformer = T5EncoderModel.from_pretrained(version)
+        self.transformer = T5EncoderModel.from_pretrained(
+            version, dtype=torch.float32
+        )
         self.device = device
         self.max_length = max_length  # TODO: typical value?
         if freeze:
@@ -145,7 +147,9 @@ class FrozenCLIPEmbedder(AbstractEncoder):
         super().__init__()
         assert layer in self.LAYERS
         self.tokenizer = CLIPTokenizer.from_pretrained(version)
-        self.transformer = CLIPTextModel.from_pretrained(version)
+        self.transformer = CLIPTextModel.from_pretrained(
+            version, dtype=torch.float32
+        )
         self.device = device
         self.max_length = max_length
         if freeze:
@@ -229,9 +233,13 @@ class FrozenCLIPEmbedderT3(AbstractEncoder):
     ):
         super().__init__()
         self.tokenizer = CLIPTokenizer.from_pretrained(version)
-        self.transformer = CLIPTextModel.from_pretrained(version)
+        self.transformer = CLIPTextModel.from_pretrained(
+            version, dtype=torch.float32
+        )
         if use_vision:
-            self.vit = CLIPVisionModelWithProjection.from_pretrained(version)
+            self.vit = CLIPVisionModelWithProjection.from_pretrained(
+                version, dtype=torch.float32
+            )
             self.processor = AutoProcessor.from_pretrained(version)
         self.device = device
         self.max_length = max_length
